Round 5: Fix compilation errors - imports, type mismatches, API fixes
Fixed 57 errors across 24 files:
- border import: androidx.compose.ui.draw.border → androidx.compose.foundation.border
- Color references: Color.White → ComposeColor.White (alias fixes)
- Type conversions: Double.toFloat(), Int.toDouble() for math operations
- Modifier fixes: horizontalArrangement → verticalArrangement for Column
- AndroidView import: androidx.compose.ui.view → androidx.compose.ui.platform
- Function renames: superscript → superscriptGCD, QRScanHistoryItem → ScanHistoryItem
- pow(): kotlin.math.pow(a,b) → a.pow(b) (extension function)
- roundToLong → kotlin.math.round(x).toLong()
- Typos: .md/.dm → .dp, duplicate modifier parameters
- Text composable: minWidth → Modifier.width()
- Added ExperimentalLayoutApi for FlowRow usage
- Added missing imports: URL, background, weight, isPrimeLocal helper

+++ app/src/main/java/com/dannyk/toolbox/ui/screens/tools/image/ContrastCheckerScreen.kt
@@ -202,7 +202,7 @@ fun ContrastCheckerScreen(navController: NavHostController) {
                     level = "AA",
                     textType = "Normal Text (<18px or <14px bold)",
                     requiredRatio = "4.5:1",
-                    actualRatio = contrastRatio,
+                    actualRatio = contrastRatio.toFloat(),
                     passes = contrastRatio >= 4.5
                 )
 
@@ -211,7 +211,7 @@ fun ContrastCheckerScreen(navController: NavHostController) {
                     level = "AA",
                     textType = "Large Text (≥18px or ≥14px bold)",
                     requiredRatio = "3:1",
-                    actualRatio = contrastRatio,
+                    actualRatio = contrastRatio.toFloat(),
                     passes = contrastRatio >= 3.0
                 )
 
@@ -220,7 +220,7 @@ fun ContrastCheckerScreen(navController: NavHostController) {
                     level = "AAA",
                     textType = "Normal Text (<18px or <14px bold)",
                     requiredRatio = "7:1",
-                    actualRatio = contrastRatio,
+                    actualRatio = contrastRatio.toFloat(),
                     passes = contrastRatio >= 7.0
                 )
 
@@ -229,7 +229,7 @@ fun ContrastCheckerScreen(navController: NavHostController) {
                     level = "AAA",
                     textType = "Large Text (≥18px or ≥14px bold)",
                     requiredRatio = "4.5:1",
-                    actualRatio = contrastRatio,
+                    actualRatio = contrastRatio.toFloat(),
                     passes = contrastRatio >= 4.5
                 )
             }
